backend/requirements.txt
Django==4.2.1
djangorestframework==3.14.1
django-cors-headers==4.0.1
djangorestframework-simplejwt==5.0.1
requests==2.31.0
pillow==9.5.0
